requirements.txt
beautifulsoup4==4.9.1
boto3==1.14.57
botocore==1.17.57
celery==5.0.2
Django==3.1.1
django-celery-beat==2.2.0
django-celery-results==2.0.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-elasticsearch-dsl==7.1.4
django-environ==0.4.5
django-filter==2.3.0
django-model-utils==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.8.0
django-storages
drf-yasg==1.17.1
elasticsearch==7.9.1
elasticsearch-dsl==7.2.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
redis==3.5.3
pycurl==7.43.0.1
mailjet_rest==1.3.4
click==7.1.1
openpyxl==3.0.10
PyJWT==2.1.0